testing/python/image_annotation/pyproject.toml
[project]
name = "image-annotation"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "numpy",
    "pandas",
    "opencv-python-headless",
    "huggingface-hub",
    "beartype",
    "pillow>=12.1.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "onnx>=1.20.1",
    "onnx2torch>=1.5.15",
    "opencv-python>=4.13.0.92",
    "transformers>=5.3.0",
    "timm>=1.0.25",
    "onnxruntime-migraphx",
]

[tool.uv.sources]
torch = { index = "pytorch-rocm" }
torchvision = { index = "pytorch-rocm" }
onnxruntime-migraphx = { path = "../../../../onnxruntime/build/Linux/Release/dist/onnxruntime_migraphx-1.25.0-cp314-cp314-linux_x86_64.whl" }

[tool.uv]
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm7.0"

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
